geotags/serializer.py
from rest_framework import serializers

from geotags.models import Point, Message

class PointSerializer(serializers.ModelSerializer):
    class Meta:
        model = Point
        fields = '__all__'
        read_only_fields = ('created_at',)

class MessageSerializer(serializers.ModelSerializer):
    class Meta:
        model = Message
        fields = '__all__'
        read_only_fields = ('created_at',)
