pyproject.toml
[project]
name = "geotags"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "django==5",
    "djangorestframework>=3.16.1",
]
